relocated api-service folder
