+++ src/shared/index.js
@@ -1,1 +1,2 @@
 export { useActions } from './hooks/index.js';
+export { ApiService } from './api-service/index.js';
